refactor: 컴포넌트 타입 정리 [TM-125]
