+++ src/shared/components/dropdown/plate/Dropdown.tsx
@@ -1,13 +1,9 @@
-import { ReactNode, useRef } from 'react';
+import { useRef } from 'react';
 import styled from 'styled-components';
 import DropdownTrigger from './Trigger';
 import DropdownMenu from './Menu';
 import useClickOutside from '@/shared/hooks/action/useClickOutstide';
-
-interface IDropdown {
-  children: ReactNode;
-  setIsOpen: React.Dispatch<React.SetStateAction<boolean>>;
-}
+import { IDropdown } from '@/shared/types';
 
 /**
  * 드롭다운 트리거 및 메뉴를 위한 컨테이너를 제공하는 드롭다운 컴포넌트입니다.
@@ -16,18 +12,16 @@ interface IDropdown {
  * @param {React.Dispatch<React.SetStateAction<boolean>>} setIsOpen - 드롭다운의 열림 상태를 설정하는 함수입니다.
  */
 
-const Dropdown = ({ children, setIsOpen }: IDropdown) => {
+export default function Dropdown({ children, setIsOpen }: IDropdown) {
   const dropdownRef = useRef<HTMLDivElement>(null);
 
   useClickOutside(dropdownRef, setIsOpen);
 
   return <DropdownContainer ref={dropdownRef}>{children}</DropdownContainer>;
-};
+}
 
 Dropdown.Trigger = DropdownTrigger;
 Dropdown.Menu = DropdownMenu;
-
-export { Dropdown };
 
 const DropdownContainer = styled.div`
   position: relative;
