--- src/shared/types/components/skeleton.types.ts
@@ -1,6 +1,0 @@
-interface ISkeletonProps {
-  width: number;
-  height: number;
-}
-
-export type { ISkeletonProps };
